rename modals.js to model.js
