+++ client/js/config/models.js
@@ -2,7 +2,8 @@ import AsyncStorage from '@react-native-community/async-storage';
 
 export const queryViewer = async () => {
   try {
-    return await AsyncStorage.getItem('user');
+    const value = await AsyncStorage.getItem('user');
+    return JSON.parse(value);
   } catch (error) {
     console.log(error);
     return error;
